Métodos de construção de queries refatorados
